requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
langchain==0.2.16
langchain-core==0.2.52
langchain-community==0.2.17
langchain-openai==0.1.25
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pypdf==4.2.0
python-docx==1.1.0
docx2txt==0.8
